Split API up into seperate files to make it a little cleaner
