+++ updater/updater.go
@@ -1,13 +1,13 @@
 package updater
 
 import(
-  "github.com/yosmudge/graphatmo/api"
+  "github.com/yosmudge/graphatmo/netatmo"
   "github.com/yosmudge/graphatmo/config"
 )
 
 type Updater struct{
   Config      *config.Config
-  Api         api.Api
+  Api         netatmo.Api
   NoSend      bool
 }
 
@@ -15,7 +15,7 @@ func New(c *config.Config, noSend bool) *Updater{
   u := &Updater{}
   u.Config = c
 
-  u.Api, _ = api.Create(*c)
+  u.Api, _ = netatmo.Create(*c)
 
   u.NoSend = noSend
